Placeholder icons for social (#458)
* Add social icons to RSS articles
* Add node 8 to test runs. Add more fail scenarios to stream test
* Increase timeout to 60s to enable test to pass on node 8 and 9

+++ app/src/AppRouter.js
@@ -40,7 +40,7 @@ class AppRouter extends Component {
 					});
 				})
 				.catch(err => {
-					if (err.response.status === 401 || err.response.status === 404) {
+					if (err.response && (err.response.status === 401 || err.response.status === 404)) {
 						localStorage.clear();
 						window.location = '/';
 					}
